backend/python/requirements.txt
# FastAPI 和相关依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
httpx>=0.25.0

# 数据库
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0  # SQLite 异步驱动
asyncpg>=0.29.0  # PostgreSQL 异步驱动（可选）

# 配置管理
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# 文件处理
PyMuPDF>=1.23.0  # PDF 处理
python-docx>=1.1.0  # Word 处理
openpyxl>=3.1.0  # Excel 处理
python-pptx>=0.6.23  # PPT 处理
Pillow>=10.0.0  # 图片处理
pandas>=2.0.0  # 数据处理

# OpenAI SDK（用于 LLM 服务）
openai>=1.0.0

# LangGraph 和 LangChain（用于 Agent 工具调用）
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# 其他工具
python-multipart>=0.0.6  # 文件上传支持
